move job/status to job.go
